Remove the global modelfilename and declare the datafilename as string.
As far as possible we should avoid having global variables. The modelfilename
is currently passed to parse_model(), and it's not used elsewhere, so it can
be removed.

The datafilename is used in parse_model() through the GlobalVariable class:
it could be removed as well if we passed it as an argument to parse_model(),
but eventually when we fix and reactivate the parse_data() code we will not
use it anymore within parse_model().