Stochastic models shouldn't need to use Exp() in the objective.
For stochastic models, the objective should be interpreted as an expectation.
This implements this interpretation, so that we do not need to use Exp() in
the objective to produce the correct solution.

We now report an error if we use Exp() inside an objective. This behaviour
is controlled by the STRICT_ERROR_FOR_EXP_INSIDE_OBJECTIVE #define.